- polishing in type declarations
