+++ converter.go
@@ -1,26 +1,29 @@
 package roadrunner_temporal
 
 import (
-	"errors"
 	"fmt"
+	"github.com/spiral/errors"
 
 	jsoniter "github.com/json-iterator/go"
 	commonpb "go.temporal.io/api/common/v1"
 	"go.temporal.io/sdk/converter"
 )
 
-type DataConverter struct {
-	fallback converter.DataConverter
-}
+type (
+	DataConverter struct {
+		fallback converter.DataConverter
+	}
 
-type RRPayload struct {
-	Data []interface{} `json:"data"`
-}
+	RRPayload struct {
+		Data []interface{} `json:"data"`
+	}
+)
 
 func NewDataConverter(fallback converter.DataConverter) converter.DataConverter {
 	return &DataConverter{fallback: fallback}
 }
 
+// ToPayloads converts a list of values.
 func (r *DataConverter) ToPayloads(values ...interface{}) (*commonpb.Payloads, error) {
 	res := &commonpb.Payloads{}
 	for i := 0; i < len(values); i++ {
@@ -45,17 +48,20 @@ func (r *DataConverter) ToPayloads(values ...interface{}) (*commonpb.Payloads, e
 	return res, nil
 }
 
+// ToPayload converts single value to payload.
 func (r *DataConverter) ToPayload(value interface{}) (*commonpb.Payload, error) {
 	return r.fallback.ToPayload(value)
 }
 
+// FromPayloads converts to a list of values of different types.
+// Useful for deserializing arguments of function invocations.
 func (r *DataConverter) FromPayloads(payloads *commonpb.Payloads, valuePtrs ...interface{}) error {
 	if payloads == nil {
 		return nil
 	}
 
 	if len(valuePtrs) < 1 {
-		return errors.New("valuePTRs len less than 0")
+		return errors.E("valuePTRs len less than 0")
 	}
 
 	for i := 0; i < len(payloads.Payloads); i++ {
@@ -68,6 +74,7 @@ func (r *DataConverter) FromPayloads(payloads *commonpb.Payloads, valuePtrs ...i
 	return nil
 }
 
+// FromPayload converts single value from payload.
 func (r *DataConverter) FromPayload(payload *commonpb.Payload, valuePtr interface{}) error {
 	switch res := valuePtr.(type) {
 	case *RRPayload:
@@ -75,8 +82,7 @@ func (r *DataConverter) FromPayload(payload *commonpb.Payload, valuePtr interfac
 		// TODO: BYPASS MARSHAL AND SEND IT AS IT IS
 		err := jsoniter.Unmarshal(payload.GetData(), &data)
 		if err != nil {
-			return fmt.Errorf(
-				"unable to decode argument: %T, with error: %v", valuePtr, err)
+			return errors.E("unable to decode argument: %T, with error: %v", valuePtr, err)
 		}
 		res.Data = append(res.Data, data)
 	default:
@@ -85,10 +91,67 @@ func (r *DataConverter) FromPayload(payload *commonpb.Payload, valuePtr interfac
 	return nil
 }
 
+// ToString converts payload object into human readable string.
 func (r *DataConverter) ToString(input *commonpb.Payload) string {
 	return r.fallback.ToString(input)
 }
 
+// ToStrings converts payloads object into human readable strings.
 func (r *DataConverter) ToStrings(input *commonpb.Payloads) []string {
 	return r.fallback.ToStrings(input)
+}
+
+// TODO: OPTIMIZE
+func FromPayloads(dc converter.DataConverter, payloads *commonpb.Payloads, values *[]jsoniter.RawMessage) error {
+	if payloads == nil {
+		return nil
+
+	}
+	*values = make([]jsoniter.RawMessage, 0, len(payloads.Payloads))
+
+	payload := RRPayload{}
+	if err := dc.FromPayloads(payloads, &payload); err != nil {
+		return errors.E(errors.Op("decodePayload"), err)
+	}
+
+	// this is double serialization, we should remove it
+	for _, value := range payload.Data {
+		data, err := jsoniter.Marshal(value)
+		if err != nil {
+			return errors.E(errors.Op("encodeResult"), err)
+		}
+
+		*values = append(*values, data)
+	}
+
+	return nil
+}
+
+// TODO: OPTIMIZE
+func ToPayloads(dc converter.DataConverter, values []jsoniter.RawMessage, result *commonpb.Payloads) error {
+	if len(values) == 0 {
+		return nil
+	}
+
+	*result = commonpb.Payloads{
+		Payloads: make([]*commonpb.Payload, 0, len(values)),
+	}
+
+	for _, value := range values {
+		var raw interface{}
+
+		err := jsoniter.Unmarshal(value, &raw)
+		if err != nil {
+			return err
+		}
+
+		out, err := dc.ToPayload(raw)
+		if err != nil {
+			return err
+		}
+
+		result.Payloads = append(result.Payloads, out)
+	}
+
+	return nil
 }
